refactor(songbook): fold sharing into the edit modal
Move the public/visibility toggle and the share link + copy action
into SongbookFormModal so the edit modal is the single place to
manage a songbook's settings. Remove the standalone Share modal and
its toolbar action from the songbook detail view.
